+++ hunmin-adapter/src/main/kotlin/com/adapter/auth/config/EncryptConfig.kt
@@ -1,4 +1,4 @@
-package com.adapter.auth
+package com.adapter.auth.config
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
